+++ src/lib.rs
@@ -16,11 +16,12 @@
 //! Both low-level and high-level APIs are provided. The low-level API requires
 //! manually updating an encrypter with chunks of plaintext and receiving
 //! ciphertext, or updating a decrypter with chunks of ciphertext and receiving
-//! plaintext. The current high-level API implements Rust's [`Read`] and
-//! [`Write`] traits to provide a simple to use way to read and write files.
+//! plaintext. High-level APIs are provided for Rust's [`Read`], [`BufRead`],
+//! and [`Write`] traits.
 //!
 //! [saltlick spec]: https://github.com/saltlick-crypto/saltlick-spec
 //! [`Read`]: https://doc.rust-lang.org/std/io/trait.Read.html
+//! [`BufRead`]: https://doc.rust-lang.org/std/io/trait.BufRead.html
 //! [`Write`]: https://doc.rust-lang.org/std/io/trait.Write.html
 //!
 //! # Usage
@@ -86,10 +87,10 @@ pub mod crypter;
 pub mod read;
 pub mod write;
 
+mod commonio;
 mod error;
 mod key;
 mod state;
-mod sync;
 mod version;
 
 pub use self::{