+++ src/write.rs
@@ -6,39 +6,154 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use crate::{crypter::Encrypter, key::PublicKey};
+//! Wrapper types over [`Write`] objects.
+//!
+//! [`Write`]: https://doc.rust-lang.org/std/io/trait.Write.html
+
+use crate::{
+    commonio::{self, Buffer},
+    crypter::{Decrypter, Encrypter, DEFAULT_BLOCK_SIZE, MIN_BLOCK_SIZE},
+    key::{PublicKey, SecretKey},
+};
 use std::{
     cmp,
     io::{self, Write},
 };
 
-const MIN_BUF_SIZE: usize = 1024;
-const DEFAULT_BUF_SIZE: usize = 32 * 1024;
-
-/// Wraps an underlying writer with encryption using the saltlick format.
+/// Wraps an underlying writer with decryption using the saltlick format.
 ///
-/// Wraps a writer and returns a type that also implements `Write`. Any data
-/// written to the returned type will be encrypted with the provided public key
-/// and written to the underlying writer.
+/// Wraps a writer that implements [`Write`] and returns a type that also
+/// implements [`Write`]. Any data written to the returned type will be
+/// decrypted with the provided secret key and written to the underlying
+/// writer.  The public key is also checked to make sure this encrypted data is
+/// for the public/secret key pair provided.
+///
+/// Any data written after the encrypted stream is complete will be discarded.
 ///
 /// # Note
 ///
 /// If writing fails during the stream, the entire stream must be discarded and
 /// started over, as the underlying crypto prevents restarting.
+///
+/// [`Write`]: https://doc.rust-lang.org/std/io/trait.Write.html
+#[derive(Debug)]
+pub struct SaltlickDecrypter<W: Write> {
+    buffer: Buffer,
+    decrypter: Decrypter,
+    inner: Option<W>,
+}
+
+impl<W: Write> SaltlickDecrypter<W> {
+    /// Create a new decryption layer over `writer` using `secret_key` and `public_key`.
+    pub fn new(public_key: PublicKey, secret_key: SecretKey, writer: W) -> SaltlickDecrypter<W> {
+        Self::with_capacity(DEFAULT_BLOCK_SIZE, public_key, secret_key, writer)
+    }
+
+    /// Create a new decryption layer over `writer`, using `lookup_fn` to match
+    /// the stream's `public_key` to its `secret_key`.
+    pub fn new_deferred<F>(writer: W, lookup_fn: F) -> SaltlickDecrypter<W>
+    where
+        F: FnOnce(&PublicKey) -> Option<SecretKey> + 'static,
+    {
+        Self::deferred_with_capacity(DEFAULT_BLOCK_SIZE, writer, lookup_fn)
+    }
+
+    /// Create a new decryption layer over `writer` using `secret_key` and
+    /// `public_key` with the provided buffer `capacity`.
+    pub fn with_capacity(
+        capacity: usize,
+        public_key: PublicKey,
+        secret_key: SecretKey,
+        writer: W,
+    ) -> SaltlickDecrypter<W> {
+        let capacity = cmp::max(capacity, MIN_BLOCK_SIZE);
+        SaltlickDecrypter {
+            buffer: Buffer::new(capacity),
+            decrypter: Decrypter::new(public_key, secret_key),
+            inner: Some(writer),
+        }
+    }
+
+    /// Create a new decryption layer over `writer`, using `lookup_fn` to match
+    /// the stream's `public_key` to its `secret_key` with the provided buffer
+    /// capacity.
+    pub fn deferred_with_capacity<F>(
+        capacity: usize,
+        writer: W,
+        lookup_fn: F,
+    ) -> SaltlickDecrypter<W>
+    where
+        F: FnOnce(&PublicKey) -> Option<SecretKey> + 'static,
+    {
+        let capacity = cmp::max(capacity, MIN_BLOCK_SIZE);
+        SaltlickDecrypter {
+            buffer: Buffer::new(capacity),
+            decrypter: Decrypter::new_deferred(lookup_fn),
+            inner: Some(writer),
+        }
+    }
+
+    /// Write any remaining plaintext to the stream and verify the stream is
+    /// finished.
+    ///
+    /// This will also be done automatically if the `SaltlickDecrypter` is
+    /// dropped, but any errors will be silently discarded, which could mean
+    /// the decrypted output is not properly finalized and therefore invalid.
+    pub fn finalize(mut self) -> Result<W, io::Error> {
+        let writer = self.inner.as_mut().expect("inner writer missing");
+        commonio::write_finalized(writer, &mut self.decrypter, &mut self.buffer)?;
+        let inner = self.inner.take().expect("inner writer missing");
+        Ok(inner)
+    }
+}
+
+impl<W: Write> Write for SaltlickDecrypter<W> {
+    fn write(&mut self, input: &[u8]) -> io::Result<usize> {
+        let writer = self.inner.as_mut().expect("inner writer missing");
+        commonio::write(writer, &mut self.decrypter, &mut self.buffer, input)
+    }
+
+    fn flush(&mut self) -> io::Result<()> {
+        let writer = self.inner.as_mut().expect("inner writer missing");
+        self.buffer.flush(writer)?;
+        writer.flush()
+    }
+}
+
+impl<W: Write> Drop for SaltlickDecrypter<W> {
+    fn drop(&mut self) {
+        if self.inner.is_some() && !self.buffer.panicked() {
+            let writer = self.inner.as_mut().unwrap();
+            let _ = self.buffer.flush(writer);
+            let _ = commonio::write_finalized(writer, &mut self.decrypter, &mut self.buffer);
+        }
+    }
+}
+
+/// Wraps an underlying writer with encryption using the saltlick format.
+///
+/// Wraps a writer that implements [`Write`] and returns a type that also
+/// implements [`Write`].  Any data written to the returned type will be
+/// encrypted with the provided public key and written to the underlying
+/// writer.
+///
+/// # Note
+///
+/// If writing fails during the stream, the entire stream must be discarded and
+/// started over, as the underlying crypto prevents restarting.
+///
+/// [`Write`]: https://doc.rust-lang.org/std/io/trait.Write.html
 #[derive(Debug)]
 pub struct SaltlickEncrypter<W: Write> {
-    available: usize,
-    ciphertext: Box<[u8]>,
-    consumed: usize,
+    buffer: Buffer,
     encrypter: Encrypter,
     inner: Option<W>,
-    panicked: bool,
 }
 
 impl<W: Write> SaltlickEncrypter<W> {
     /// Create a new encryption layer over `writer` using `public_key`.
     pub fn new(public_key: PublicKey, writer: W) -> SaltlickEncrypter<W> {
-        SaltlickEncrypter::with_capacity(DEFAULT_BUF_SIZE, public_key, writer)
+        SaltlickEncrypter::with_capacity(DEFAULT_BLOCK_SIZE, public_key, writer)
     }
 
     /// Create a new encryption layer over `writer` using `public_key` with the
@@ -48,19 +163,16 @@ impl<W: Write> SaltlickEncrypter<W> {
         public_key: PublicKey,
         writer: W,
     ) -> SaltlickEncrypter<W> {
-        let capacity = cmp::max(capacity, MIN_BUF_SIZE);
+        let capacity = cmp::max(capacity, MIN_BLOCK_SIZE);
         SaltlickEncrypter {
-            available: 0,
-            ciphertext: vec![0u8; capacity].into_boxed_slice(),
-            consumed: 0,
+            buffer: Buffer::new(capacity),
             encrypter: Encrypter::new(public_key),
             inner: Some(writer),
-            panicked: false,
         }
     }
 
     /// Set the block size for the underlying encrypter.
-    pub fn block_size(&mut self, block_size: usize) {
+    pub fn set_block_size(&mut self, block_size: usize) {
         self.encrypter.set_block_size(block_size);
     }
 
@@ -70,78 +182,158 @@ impl<W: Write> SaltlickEncrypter<W> {
     /// dropped, but any errors will be silently discarded, which could mean
     /// the encrypted output is not properly finalized and therefore invalid.
     pub fn finalize(mut self) -> Result<W, io::Error> {
-        while self.encrypter.is_not_finalized() {
-            self.flush_buf()?;
-            let (_, wr) = self.encrypter.update(&[], &mut self.ciphertext, true)?;
-            self.available = wr;
-            self.consumed = 0;
-        }
-        self.flush()?;
+        let writer = self.inner.as_mut().expect("inner writer missing");
+        commonio::write_finalized(writer, &mut self.encrypter, &mut self.buffer)?;
         let inner = self.inner.take().expect("inner writer missing");
         Ok(inner)
-    }
-
-    fn flush_buf(&mut self) -> io::Result<()> {
-        while self.ciphertext_len() > 0 {
-            self.panicked = true;
-            let writer = self.inner.as_mut().expect("inner writer missing");
-            let res = writer.write(&self.ciphertext[self.consumed..self.available]);
-            self.panicked = false;
-            match res {
-                Ok(0) => {
-                    return Err(io::Error::new(
-                        io::ErrorKind::WriteZero,
-                        "failed to write the buffered data",
-                    ));
-                }
-                Ok(n) => self.consumed += n,
-                Err(ref e) if e.kind() == io::ErrorKind::Interrupted => {}
-                Err(e) => return Err(e),
-            }
-        }
-        Ok(())
-    }
-
-    fn ciphertext_len(&self) -> usize {
-        self.available - self.consumed
     }
 }
 
 impl<W: Write> Write for SaltlickEncrypter<W> {
     fn write(&mut self, input: &[u8]) -> io::Result<usize> {
-        // All current ciphertext needs to be flushed before writing anything
-        // new since we give `update` the whole buffer.
-        self.flush_buf()?;
-
-        // Returning a zero write is an error, so keep updating until we read
-        // some input, block, or error.
-        let mut last_rd = 0;
-        while last_rd == 0 {
-            let (rd, wr) = self.encrypter.update(input, &mut self.ciphertext, false)?;
-            self.available = wr;
-            self.consumed = 0;
-            last_rd = rd;
-            self.flush_buf()?;
-        }
-        Ok(last_rd)
+        let writer = self.inner.as_mut().expect("inner writer missing");
+        commonio::write(writer, &mut self.encrypter, &mut self.buffer, input)
     }
 
     fn flush(&mut self) -> io::Result<()> {
-        self.flush_buf()?;
-        self.inner.as_mut().expect("inner writer missing").flush()
+        let writer = self.inner.as_mut().expect("inner writer missing");
+        self.buffer.flush(writer)?;
+        writer.flush()
     }
 }
 
 impl<W: Write> Drop for SaltlickEncrypter<W> {
     fn drop(&mut self) {
-        if self.inner.is_some() && !self.panicked {
-            let (_, wr) = self
-                .encrypter
-                .update(&[] as &[u8], &mut self.ciphertext, true)
-                .unwrap_or((0, 0));
-            self.available = wr;
-            self.consumed = 0;
-            let _ = self.flush();
+        if self.inner.is_some() && !self.buffer.panicked() {
+            let writer = self.inner.as_mut().unwrap();
+            let _ = self.buffer.flush(writer);
+            let _ = commonio::write_finalized(writer, &mut self.encrypter, &mut self.buffer);
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{SaltlickDecrypter, SaltlickEncrypter};
+    use crate::key::gen_keypair;
+    use rand::{RngCore, SeedableRng};
+    use rand_xorshift::XorShiftRng;
+    use std::{cmp, io::Write, iter};
+
+    fn random_bytes(seed: u64, size: usize) -> Box<[u8]> {
+        let mut rng = XorShiftRng::seed_from_u64(seed);
+        let mut bytes = vec![0u8; size];
+        rng.fill_bytes(&mut bytes);
+        bytes.into_boxed_slice()
+    }
+
+    #[test]
+    fn single_write_test() {
+        for size in &[
+            1,
+            10 * 1024,
+            32 * 1024,
+            100 * 1024,
+            200 * 1024,
+            10 * 1024 * 1024,
+        ] {
+            let random_data = random_bytes(0, *size);
+            let (public_key, secret_key) = gen_keypair();
+            let decrypter = SaltlickDecrypter::new_deferred(Vec::new(), |_| Some(secret_key));
+            let mut encrypter = SaltlickEncrypter::new(public_key, decrypter);
+            encrypter.write_all(&random_data[..]).unwrap();
+            let decrypter = encrypter.finalize().unwrap();
+            let output = decrypter.finalize().unwrap();
+            assert_eq!(&random_data[..], &output[..]);
+        }
+    }
+
+    #[test]
+    fn multiple_write_test() {
+        for size in &[
+            1,
+            10 * 1024,
+            32 * 1024,
+            100 * 1024,
+            200 * 1024,
+            10 * 1024 * 1024,
+        ] {
+            let random_data = random_bytes(0, *size);
+            let (public_key, secret_key) = gen_keypair();
+            let decrypter = SaltlickDecrypter::new_deferred(Vec::new(), |_| Some(secret_key));
+            let mut encrypter = SaltlickEncrypter::new(public_key, decrypter);
+            encrypter.set_block_size(16 * 1024);
+
+            // Take increasing chunks so we're varying chunk size.
+            let mut written = 0;
+            for take in iter::successors(Some(1usize), |n| Some(n + 7)) {
+                let end = cmp::min(written + take, *size);
+                encrypter.write_all(&random_data[written..end]).unwrap();
+                encrypter.flush().unwrap();
+                written += take;
+                if written >= *size {
+                    break;
+                }
+            }
+
+            let decrypter = encrypter.finalize().unwrap();
+            let output = decrypter.finalize().unwrap();
+            assert_eq!(&random_data[..], &output[..]);
+        }
+    }
+
+    #[test]
+    fn drop_flush_test() {
+        for size in &[
+            1,
+            10 * 1024,
+            32 * 1024,
+            100 * 1024,
+            200 * 1024,
+            10 * 1024 * 1024,
+        ] {
+            let random_data = random_bytes(0, *size);
+            let (public_key, secret_key) = gen_keypair();
+            let mut output = Vec::new();
+            {
+                let decrypter = SaltlickDecrypter::new(public_key.clone(), secret_key, &mut output);
+                let mut encrypter = SaltlickEncrypter::new(public_key.clone(), decrypter);
+                encrypter.write_all(&random_data[..]).unwrap();
+                // Dropping encrypter/decrypter here should finalize content in `output`
+            }
+            assert_eq!(&random_data[..], &output[..]);
+        }
+    }
+
+    #[test]
+    fn corrupt_value_test() {
+        let random_data = random_bytes(0, 100 * 1024);
+        let (public_key, secret_key) = gen_keypair();
+        let mut encrypter = SaltlickEncrypter::new(public_key.clone(), Vec::new());
+        encrypter.write_all(&random_data[..]).unwrap();
+        let mut ciphertext = encrypter.finalize().unwrap();
+
+        // Inject a single bad byte near the end of the stream
+        let index = ciphertext.len() - 5;
+        ciphertext[index] = ciphertext[index].wrapping_add(1);
+
+        let mut decrypter = SaltlickDecrypter::new(public_key, secret_key, Vec::new());
+        assert!(decrypter.write_all(&ciphertext[..]).is_err());
+    }
+
+    #[test]
+    fn incomplete_stream_test() {
+        let random_data = random_bytes(0, 100 * 1024);
+        let (public_key, secret_key) = gen_keypair();
+        let mut encrypter = SaltlickEncrypter::new(public_key.clone(), Vec::new());
+        encrypter.write_all(&random_data[..]).unwrap();
+        let mut ciphertext = encrypter.finalize().unwrap();
+
+        // Remove a few bytes from the end
+        ciphertext.resize(ciphertext.len() - 5, 0);
+
+        let mut decrypter = SaltlickDecrypter::new(public_key, secret_key, Vec::new());
+        decrypter.write_all(&ciphertext[..]).unwrap();
+        assert!(decrypter.finalize().is_err());
     }
 }